refactor: extract Figure, GridSpec, SubplotSpec into figure/ subpackage
